Highlight inlabel: results
When using the inlabel: keyword only a filter was attached, so no
highlighting of the chosen label was performed. Add the filter to
nonTextHighlightQueries to make that possible.

CirrusSearch additionally requires patch Ic13f82282, otherwise only
the label can be highlighted an additional parts of the query are
not considered.

Bug: T217809
Depends-On: Ic13f822826c46c6ef6417a40fa846524e49173eb
Change-Id: Ic1ae01d10076993a79dd0168e84089560ea4ec54