Update HasDataForLangFeature for caption move to descriptions
In I0463fb936 captions were moved from the labels field to the
descriptions field. Update HasDataForLangFeature to match
the new definition.

Additionally add '*' support to hasdescription, to match the
previous functionality of hascaption when aliased to haslabel.
The current implementation will have very low performance, but
the query volume is exceptionally low and this can be revisited
with an appropriately indexed field if necessary later.

Bug: T226722
Change-Id: I91dfa3f0c5eba87a180c068d6edd2224c73064b0