tests: Remove empty placeholder for selenium tests
This test is not doing anything useful. And over a year since
T223789 was filed, still nothing has been written. From a quick
glance it seems to not be a priority since for the most part it
is not user-facing and the team at WMDE is not prioritising this
currently.

Given we spend a full minute running through this no-op on every
MW patch and backport, remove it for now.

Bug: T223789
Change-Id: Icf029f53bf4145d0e77cb6bffdfba73f9585ee1f