Allow dangling commas in whitespace-separated license config
It's documented (in i18n message) to accept dangling comma's
(which make it much easier to append changes without breaking
things), so this needs to work.

This will only strip whitespace if there is no `<something>|`
detected after the whitespace. This will preserve the newline
that should preceed any license config.

Bug: T257938
Change-Id: I3c950bb9c2d57c27580eb9ae9c669ddd6767b067