Remove support for prefixed entity IDs and foreignRepositories
The foreignRepositories approach was ultimately not used for Structured
Data on Commons – instead, there are multiple entitySources, and items
and properties always come from the Wikidata source, which means they’re
not prefixed. As we prepare to remove the last vestiges of support for
foreignRepositories from Wikibase, remove it here as well.

In theory, EntityId objects can still contain a repository name, which
may appear in getSerialization() calls, but we don’t expect that to
happen, so I didn’t add any related code in StatementsField (just
removed the tests for that in StatementsFieldTest).

Bug: T286767
Change-Id: I82582576193d13d983717347208e131b7d342ff1