Apply minimum_should_match for elastic 7
In elasticsearch 7.x the minimum_should_match value defaults to 0,
instead of 1, when used in a filter context. Due to the way cirrus
'random' sorting works all queries can be used in a filter context.

Set the value to 0 if there is an explicit filter clause that
is expected to provide filtering, otherwise set minimum_should_match
to 1 to ensure filtering is applied as expected.

Bug: T218994
Change-Id: Ib267c61b555127821b7d9451bba999789da535f9