Encode empty label/description fields as null
When importing cirrus documents into strictly typed systems we found
that the labels and descriptions fields could be found as either a map
or an empty array. These downstream systems will be simpler if we can
use consistent typing from the start. Encoding an empty map would be a
bit error prone, we would have to pass \stdClass through in the right
places so json_encode turns it into a map. Instead to make things less
surprising provide null for the empty field.

Bug: T322327
Change-Id: I2726b8ab80da0243112305119e6098cb51a0203f