StatementQuantityFieldTest: Don’t load entities in data provider
If we load the test entities directly in the data provider, the test
services won’t have been set up yet, and Wikibase will try to
deserialize data values based on the data types of whatever properties
the wiki actually happens to have, unrelated to the test data. This
breaks in CI and can break locally as well. Instead, make the data
provider yield plain entity IDs and load the test data only in the test
function (note that StatementsFieldTest already did this), and also
register our fake data type lookup in the service container so that it’s
used during deserialization (otherwise trying to access the real data
type lookup, when this test isn’t in the Database group, will yield an
error “Database backend disabled”).

Bug: T369357
Change-Id: I078e7f82b565b81349e34decb0f3c5b5646972cd