StatementsField: Log warning via logger
This means the warning message can be normalized, so we don’t get a
logspam-watch full of almost identical messages (with different
statement GUIDs) when someone deletes a property before removing its
uses (as happened today).

Note that StatementProviderFieldDefinitions is also created from
WikibaseMediaInfo (both constructor and newFromSettings()), so it’s
necessary for compatibility that the $logger parameter is at the end of
the parameter list and has a default null argument. The argument is
added in WikibaseMediaInfo change Ibb03ef7a9d.

I initially considered using a separate log channel here (i.e.
LoggerFactory::getInstance( 'WikibaseCirrusSearch' )), but this seems
weird if one of the loggers is created in WikibaseMediaInfo (should
WikibaseMediaInfo still hard-code the name WikibaseCirrusSearch? or
should the log channel be different in that case even though the actual
warning condition is identical?), so I decided to keep it simple and
just reuse the Wikibase log channel. (We already do this a handful of
times in WikibaseLexeme and WikibaseMediaInfo anyway.)

Change-Id: I6f89979272498fe22c43538c48d8279231fc51dc