Index plain property id into statement_keywords
When building out the statement_keywords functionality we limited
the set of properties that would be indexed as we wanted to avoid
the cardinality of all possible values. That only really applies
to the values though, the set of possible properties themselves
is reasonably limited.

By always including at least the P<n> value this allows the
`haswbstatement` keyword to cover more use cases and have less
awkward edge cases.

Bug: T372030
Change-Id: I11deb0a4e67ac76be4dfe43281d6642caa281f46