Document dependency on experimental highlighter plugin
Currently, the extension unconditionally uses the experimental
highlighter, so if the plugin isn’t installed, wbsearchentities will
return empty results. I’m not sure if this is intentional or not, but it
seems easier to document it than to make the extension use the
highlighter only if available, so let’s just add it to the README.

Also include links for both plugins instead of just leaving the reader
to search for them on their own.

Bug: T373670
Change-Id: Id704eed6fb44e404f3e3f11753c4eae56885a101
(cherry picked from commit 4d2281824bb5cae8e605340b26e6dec61976180c)