Inject ConfigFactory into LabelsProviderFieldDefinitions
(And DescriptionsProviderFieldDefinitions.)

While LabelsProviderFieldDefinitions isn’t really supposed to be part of
the public interface of the extension, it is currently used in
WikibaseMediaInfo and might soon be used in EntitySchema as well. This
means that, to correctly construct the field definitions with stemming
support, those extensions not only have to know about the
LabelsProviderFieldDefinitions class itself but also how to get the
stemming settings from the UseStemming option of the
WikibaseCirrusSearch config from the config factory. By moving this part
into LabelsProviderFieldDefinitions itself, we reduce the amount of
WikibaseCirrusSearch internals that other extensions have to deal with.

For compatibility with WikibaseMediaInfo, this change temporarily
accepts both a ConfigFactory and an array. This will be cleaned up
immediately.

Bug: T377318
Change-Id: I2d925240ba808fc16e41001ca1f6fb8dbe7a63c7