Introduce WikibaseLabelsIndexField
WikibaseLabelsIndexField is a new interface that specializes
WikibaseIndexField for fields responsible for indexing "labels".
Labels are multi-lingual text attached to entities and generally
extracted using the LabelsProvider & AliasesProvider interfaces.

By specializing a WikibaseIndexField for labels and introducing a new
method (getLabelsIndexData) that takes a LabelsProvider, re-users having
entities where labels could be extracted but that are not themselves
subclasses of EntityDocument can now populate such fields in the search
index without having to understand the internal structure of the labels
fields.

LabelsProviderFieldDefinitions will now return a list of
WikibaseLabelsIndexField such that extension willing to re-use this can
"simply" inject a LabelsProviderFieldDefinitions into their
ContentHandler.

Change-Id: Ic40ed179bb9e93b9e868874dad997af01a7af5ff