Make ElasticTermResult reusable in EntitySchema
Extract the part that uses an EntityIdParser into the new subclass
EntityElasticTermResult, so that ElasticTermResult can be used to power
EntitySchema search too.

I’m not very happy with the name EntityElasticTermResult – it’s
uncomfortably close to EntityResult, which AFAICT is a completely
separate class hierarchy – but at least this name is internal and can
easily be changed later (it’s not referenced in other extensions).

Bug: T376252
Change-Id: Ia9b69fed0bd8de4bef66c597dd7f9a89e42c10a0
Depends-On: I92d6519556a2c471100642820975fce41336eb9f