StatementsField: Check data type and data value type
If the property type was changed in such a way that old data values are
not compatible with the new property type, we shouldn’t try to call the
search formatter, as it’ll probably just crash anyway. Bail out and skip
indexing that statement instead. (Such cases should be extremely rare on
“real” wikis – even the changePropertyDataType maintenance script won’t
let you change the data type in an incompatible fashion – but there are
some examples of this on Test Wikidata, so we need to handle it to avoid
logspam in production.)

Unfortunately, injecting a DataTypeFactory into StatementsField is
complicated by the fact that other extensions also instantiate the
class. Most of them, including all non-test code, go through the
newFromSettings() method, so add a temporarily compatibility layer
there; the other tests can be skipped (see Depends-On).

Bug: T372993
Change-Id: I7e2490c25b720f3499638552154d24622f32ba64
Depends-On: I6ef5049d1a8e35e65bc67c907319d52441bc3aa5