Generalize WikibasePrefixSearcher
This is going to be used for the plain label search as well. Nothing
about it is prefix search specific, so we can just give it a more
general name, and inject the stats and syntax keys.

The original file is kept around here for compatibility with other git
repos that depend on the class. It will be removed in a follow-up.

Bug: T387954
Change-Id: I5853b807c32fc4827efb75978c947e72367ff5e1