Add InlabelSearch
This introduces a new type of entity search by label (and aliases),
which is similar to using the existing "inlabel:" full text search
keyword, but as a standalone InLabelSearch service.

The search profiles, service class and query were adapted from the
existing prefix search ones. InLabelSearch and InLabelSearchQuery are to
the new type of "in label search" what EntitySearchElastic and
LabelsCompletionQuery are to the existing prefix search.

Bug: T387954
Change-Id: I12813a14cf9ba39aa8ec050644ea05baaa3d2d09