Update InLabelQuery to support AND, OR, NOT, and "phrase matching"
The QueryParser also does some advanced parsing, such as fuzzy words and wildcards, that we aren't making use of. If we continue to not use these features, we should modify the QueryStringRegexParser to enable turning off those features.

There are some parts of the scoring query that we may want to tweak after some more thorough testing (see the TODOs in InLabelScoringVisitor). These will be done in a follow-up patch.

Bug: T389011
Depends-On: I0a8046a86bcb3ccee5b2b612daf623d698b5c422
Change-Id: If01143b9f29c81f9f6599c82ebe6e6b15b3a6681