Extract LabelsDescriptionsFieldTrait
To reduce some duplicate code between LabelsField and DescriptionsField.

Note that Phan gets unhappy if we just move the “real” methods
(undeclared property type, undeclared method __construct()), which is
why I moved the properties and constructor as well. A few remaining Phan
issues still have to be suppressed AFAICT (the “potential instantiation
of trait” is a false positive, https://github.com/phan/phan/issues/2631;
the two errors on the parent constructor calls are because we can’t
declare statically that the class must extend SearchIndexFieldDefinition
and have a NAME constant).

Change-Id: Ie2dc01274cace4e49ca29f46ce933c0a10f51bbd