WikibaseRepoEntityTypesHookHandler: Don’t reorder entity types
We need to use the search entity type definitions as the base of the
addition, and add the original entity type definitions onto it (not vice
versa), so that the search callback from the search entity type
definitions will override the original one. However, the way this was
previously implemented also meant that the resulting array would start
with the keys from the search entity type definitions, and only then any
other keys from the original definitions, potentially changing the order
of the array. This side effect is both undesirable and visible to users
in some cases, such as in the order the entity types are listed in the
scoped typeahead search. Rewrite the code to do the array addition in a
different way that preserves the order of the original array, and then
add some tests for the hook handler as well.

In WikibaseCirrusSearch, this issue actually had no effect; it’s more
important in the WikibaseLexemeCirrusSearch version of this hook
handler. However, I think it’s a good idea to fix this one as well. (See
change If2277f42cc for the WikibaseLexemeCirrusSearch version.)

Bug: T398817
Change-Id: I206534998d15b9b11fbc8b41623ad20d911969b0